Move rotations into math file
