+++ src/berreman4x4/__init__.py
@@ -5,7 +5,6 @@ from .experiment import Experiment
 from .materials import *
 from .math import lambda2E
 from .result import Result
-from .rotations import *
 from .settings import *
 from .solver import *
 from .structure import *
